+++ packages/query-core/src/__tests__/infiniteQueryBehavior.test.tsx
@@ -31,8 +31,7 @@ describe('InfiniteQueryBehavior', () => {
     })
 
     let observerResult:
-      | InfiniteQueryObserverResult<unknown, unknown>
-      | undefined
+      InfiniteQueryObserverResult<unknown, unknown> | undefined
 
     const unsubscribe = observer.subscribe((result) => {
       observerResult = result
@@ -67,8 +66,7 @@ describe('InfiniteQueryBehavior', () => {
     })
 
     let observerResult:
-      | InfiniteQueryObserverResult<unknown, unknown>
-      | undefined
+      InfiniteQueryObserverResult<unknown, unknown> | undefined
 
     const unsubscribe = observer.subscribe((result) => {
       observerResult = result
@@ -215,8 +213,7 @@ describe('InfiniteQueryBehavior', () => {
     })
 
     let observerResult:
-      | InfiniteQueryObserverResult<unknown, unknown>
-      | undefined
+      InfiniteQueryObserverResult<unknown, unknown> | undefined
 
     const unsubscribe = observer.subscribe((result) => {
       observerResult = result
@@ -266,8 +263,7 @@ describe('InfiniteQueryBehavior', () => {
     })
 
     let observerResult:
-      | InfiniteQueryObserverResult<unknown, unknown>
-      | undefined
+      InfiniteQueryObserverResult<unknown, unknown> | undefined
 
     const unsubscribe = observer.subscribe((result) => {
       observerResult = result
@@ -458,8 +454,7 @@ describe('InfiniteQueryBehavior', () => {
     })
 
     let observerResult:
-      | InfiniteQueryObserverResult<unknown, unknown>
-      | undefined
+      InfiniteQueryObserverResult<unknown, unknown> | undefined
 
     const unsubscribe = observer.subscribe((result) => {
       observerResult = result