+++ packages/query-persist-client-core/src/__tests__/persist.test.ts
@@ -106,11 +106,11 @@ describe('persist', () => {
     })
 
     it('should rethrow exceptions in `restoreClient`', async () => {
-      const consoleMock = vi
+      const consoleErrorMock = vi
         .spyOn(console, 'error')
         .mockImplementation(() => undefined)
 
-      const consoleWarn = vi
+      const consoleWarnMock = vi
         .spyOn(console, 'warn')
         .mockImplementation(() => undefined)
 
@@ -125,20 +125,20 @@ describe('persist', () => {
         }),
       ).rejects.toBe(restoreError)
 
-      expect(consoleMock).toHaveBeenCalledTimes(1)
-      expect(consoleWarn).toHaveBeenCalledTimes(1)
-      expect(consoleMock).toHaveBeenNthCalledWith(1, restoreError)
+      expect(consoleErrorMock).toHaveBeenCalledTimes(1)
+      expect(consoleWarnMock).toHaveBeenCalledTimes(1)
+      expect(consoleErrorMock).toHaveBeenNthCalledWith(1, restoreError)
 
-      consoleMock.mockRestore()
-      consoleWarn.mockRestore()
+      consoleErrorMock.mockRestore()
+      consoleWarnMock.mockRestore()
     })
 
     it('should rethrow exceptions in `removeClient` before `restoreClient`', async () => {
-      const consoleMock = vi
+      const consoleErrorMock = vi
         .spyOn(console, 'error')
         .mockImplementation(() => undefined)
 
-      const consoleWarn = vi
+      const consoleWarnMock = vi
         .spyOn(console, 'warn')
         .mockImplementation(() => undefined)
 
@@ -155,12 +155,12 @@ describe('persist', () => {
         }),
       ).rejects.toBe(removeError)
 
-      expect(consoleMock).toHaveBeenCalledTimes(1)
-      expect(consoleWarn).toHaveBeenCalledTimes(1)
-      expect(consoleMock).toHaveBeenNthCalledWith(1, restoreError)
+      expect(consoleErrorMock).toHaveBeenCalledTimes(1)
+      expect(consoleWarnMock).toHaveBeenCalledTimes(1)
+      expect(consoleErrorMock).toHaveBeenNthCalledWith(1, restoreError)
 
-      consoleMock.mockRestore()
-      consoleWarn.mockRestore()
+      consoleErrorMock.mockRestore()
+      consoleWarnMock.mockRestore()
     })
 
     it('should rethrow error in `removeClient`', async () => {