+++ packages/vue-query-devtools/src/stub.ts
@@ -1,0 +1,7 @@
+export const VueQueryDevtools = function () {
+  return null
+}
+
+export const VueQueryDevtoolsPanel = function () {
+  return null
+}