+++ packages/query-core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @tanstack/query-core
 
+## 5.103.3
+
+No changes in this release.
+
 ## 5.103.2
 
 ### Patch Changes