+++ packages/query-sync-storage-persister/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/query-sync-storage-persister
 
+## 5.103.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.103.3
+  - @tanstack/query-persist-client-core@5.103.3
+
 ## 5.103.2
 
 ### Patch Changes